Fixed issues in plotting and reloading
I modified the plotting code so it is adapted to how the metrics.csv file is saved now. Also there were some issues with how the logger saved the metrics and how resuming training managed the folders for checkpointing (it overwrote the previous files in the folder), it is fixed now and works.