Added Jaccard index, testing code
I added the Jaccard index metric and validation loss in the logging, saved hyperparameters to allow weight loading, added code to load weights and perform a forward pass to some samples. This allows testing the model and comparing the prediction with the ground truth. I also modified some little things in the utils and plotting code for better functioning.