Working Code V0
With this code, the model is capable of overfitting on an image and reconstructing it. It still does not work properly on the whole dataset, I will work on that. Right now the mask is set to -1 but it seems like it works well with mask 0 too even if it has the same value as the background. Also changed the intensity scaling in the data loading code to deal with outliers.